// packages/destination-actions/src/destinations/bigquery/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Settings {
  /**
   * Your Google Cloud project ID
   */
  projectId: string
  /**
   * The dataset to insert events into
   */
  dataset: string
  /**
   * Path to your Google Cloud key file
   */
  keyFile?: string
  /**
   * The contents of your Google Cloud key file
   */
  keyData?: string
}